% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bnMCMCResults}
\alias{bnMCMCResults}
\title{Results from a Bayesian Network MCMC algorithm on simulated data}
\format{
\subsection{\code{bnMCMCResults}}{

A list with 5 elements, each representing a different MCMC chain. Each element is
a list of data.frames describing a partition arrangement of a Bayesian Network.
}
}
\source{
Luke Duttweiler
}
\usage{
bnMCMCResults
}
\description{
Results from a Bayesian Network Metropolis-Hastings algorithm run on simulated data.
Included for examples.
}
\keyword{datasets}
