% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{eucDist}
\alias{eucDist}
\title{Euclidean Distance}
\usage{
eucDist(x, y)
}
\arguments{
\item{x}{Numeric vector or matrix.}

\item{y}{Numeric vector or matrix of same dimensions as x.}
}
\value{
Numeric, elementwise Euclidean distance between x and y.
}
\description{
Simple function to return the Euclidean distance between two objects. Acts elementwise.
}
\note{
For speed, no error handling if x and y do not have the same dimensions, take care!
}
\examples{
eucDist(c(0,0), c(1,1))
}
