% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genMCMCDiag-package.R
\docType{package}
\name{genMCMCDiag-package}
\alias{genMCMCDiag}
\alias{genMCMCDiag-package}
\title{genMCMCDiag: Generalized Convergence Diagnostics for Difficult MCMC Algorithms}
\description{
Trace plots and convergence diagnostics for Markov Chain Monte Carlo (MCMC) algorithms on highly multivariate or unordered spaces. Methods outlined in a forthcoming paper.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LukeDuttweiler/genMCMCDiag}
  \item Report bugs at \url{https://github.com/LukeDuttweiler/genMCMCDiag/issues}
}

}
\author{
\strong{Maintainer}: Luke Duttweiler \email{lduttweiler@hsph.harvard.edu} (\href{https://orcid.org/0000-0002-0467-995X}{ORCID}) [copyright holder]

}
\keyword{internal}
