% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf-likelihoodTransform.R
\name{likelihoodTransform}
\alias{likelihoodTransform}
\title{Transforms a list of mcmcObjs into a list of data.frames using the 'likelihood'
transformation}
\usage{
likelihoodTransform(mhDraws, ...)
}
\arguments{
\item{mhDraws}{A list of mcmcObjs}

\item{...}{Catches extra arguments. Not used.}
}
\value{
List of data.frames with columns 'val.1' which is non-normalized
posterior of each draw and 't' which gives the within-chain ordering of the MCMC draws.
Each data.frame is a separate chain.

@export
}
\description{
Transforms a list of mcmcObjs into a list of data.frames using the 'likelihood'
transformation
}
