% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-listLabels.R
\name{listLabels}
\alias{listLabels}
\title{Function to assign character labels to all unique objects in a list}
\usage{
listLabels(lst)
}
\arguments{
\item{lst}{A list of objects. Each object in the list should have
the same general structure}
}
\value{
A character vector of labels. Objects in lst that are identical will be assigned
the same label.
}
\description{
Function to assign character labels to all unique objects in a list
}
