% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-mcmcChain.R
\docType{class}
\name{mcmcChain-class}
\alias{mcmcChain-class}
\alias{mcmcChain}
\title{mcmcChain Class Definition}
\description{
A class for storing MCMC results from a single chain.
}
\section{Slots}{

\describe{
\item{\code{val}}{A list containing the MCMC draws from a single chain IN ORDER.}

\item{\code{Posterior}}{A vector containing the Posterior values corresponding to
the draws in val}
}}

