% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-mcmcDiagMethods.R
\docType{class}
\name{mcmcDiag-class}
\alias{mcmcDiag-class}
\alias{mcmcDiag}
\title{mcmcDiag Class Definition}
\description{
A class for storing and handling MCMC diagnostics.
}
\section{Slots}{

\describe{
\item{\code{diagnostics}}{A list containing MCMC diagnostic information.}

\item{\code{transformedDraws}}{A list containing transformed MCMC draws.}

\item{\code{call}}{A list containing information about the function call.}
}}

