% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{partitionDist}
\alias{partitionDist}
\title{Partition Distance}
\usage{
partitionDist(x, y)
}
\arguments{
\item{x}{Data.frame with columns node and partition}

\item{y}{Data.frame with columns node and partition. Same nrows as x.}
}
\value{
Numeric, Partition distance between x and y.
}
\description{
Function to return the 'Partition' distance between two objects. Used for Bayesian
Networks with the 'partition-MCMC' algorithm.
}
\note{
For speed, no error handling if x and y do not have the same dimensions. Also,
does not test to make sure x,y are data.frames of integers, take care!
}
\examples{
x <- bnMCMCResults[[1]][[1]]
y <- bnMCMCResults[[1]][[100]]
partitionDist(x, y)
}
