% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-mcmcDiagMethods.R
\name{print,mcmcDiag-method}
\alias{print,mcmcDiag-method}
\alias{print.mcmcDiag}
\title{Print Method for mcmcDiag Objects}
\usage{
\S4method{print}{mcmcDiag}(x)
}
\arguments{
\item{x}{An object of class \code{mcmcDiag}.}
}
\value{
Prints MCMC diagnostics to the console.
}
\description{
Prints MCMC diagnostics and relevant information.
}
