% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag-sgelmanRubin.R
\name{sgelmanRubin}
\alias{sgelmanRubin}
\title{Calculate Gelman-Rubin statistic for multi-chain samples of a univariate (or transformed univariate) MCMC algorith from an mcmcObj}
\usage{
sgelmanRubin(mhDraws, ...)
}
\arguments{
\item{mhDraws}{Object of mcmcObj class}

\item{...}{Catches unnecessary additional arguments}
}
\value{
Data.frame with 1 row and 2 columns. First entry gives estimated GR statistic, second gives upper 95\% limit for GR statistic.
}
\description{
Calculate Gelman-Rubin statistic for multi-chain samples of a univariate (or transformed univariate) MCMC algorith from an mcmcObj
}
