% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-mcmcDiagMethods.R
\name{show,mcmcDiag-method}
\alias{show,mcmcDiag-method}
\alias{show.mcmcDiag}
\title{Show Method for mcmcDiag Objects}
\usage{
\S4method{show}{mcmcDiag}(object)
}
\arguments{
\item{object}{An object of class \code{mcmcDiag}.}
}
\value{
Displays the content of the object.
}
\description{
Displays the content of an object of class \code{mcmcDiag}.
}
