% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf-standardTransform.R
\name{standardTransform}
\alias{standardTransform}
\title{Transforms a list of MCMC chains into a list of dataframes with no modifications to values}
\usage{
standardTransform(mhDraws, ...)
}
\arguments{
\item{mhDraws}{An list of numeric vectors}

\item{...}{Not used.}
}
\value{
A list of data.frames with rows that represent MCMC draws.Each separate data.frame is a
different chain. Data.frames have columns 'val' for the numeric draws, and 't' for the
draw. Currently, using the standard transformation on anything other than univariate
draws is not supported.
}
\description{
Transforms a list of MCMC chains into a list of dataframes with no modifications to values
}
