% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag-straceplot.R
\name{straceplot}
\alias{straceplot}
\title{Generate a traceplot of MCMC draws from an mcmcObj}
\usage{
straceplot(mhDraws, method = NULL, ...)
}
\arguments{
\item{mhDraws}{Object of class mcmcObj}

\item{method}{Character string - Name of method used to generate traceplot. Is used to generate the title of the traceplot.}

\item{...}{Catches unused arguments}
}
\value{
ggplot2 plot object showing traceplot
}
\description{
Generate a traceplot of MCMC draws from an mcmcObj
}
