% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf-sumTransform.R
\name{sumTransform}
\alias{sumTransform}
\title{Transforms MCMC draws from an mcmcObj into a data.frame using the 'sum' transformation}
\usage{
sumTransform(mhDraws, ...)
}
\arguments{
\item{mhDraws}{An object of class mcmcObj}

\item{...}{Catches extra arguments. Not used.}
}
\value{
data.frame with columns 'val.1' which is the sum of all dimensions of each MCMC
draw, and 'Posterior' which is the (non-normalized) posterior value of each MCMC draw.
}
\description{
Transforms MCMC draws from an mcmcObj into a data.frame using the 'sum' transformation
}
