% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-mcmcDiagMethods.R
\name{print.mcmcDiag}
\alias{print.mcmcDiag}
\title{Print method for mcmcDiag objects}
\usage{
\method{print}{mcmcDiag}(x)
}
\arguments{
\item{x}{Object of class mcmcDiag}
}
\value{
Invisible NULL, prints to console
}
\description{
Print method for mcmcDiag objects
}
