% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-mcmcDiagMethods.R
\name{print.mcmcDiag}
\alias{print.mcmcDiag}
\title{mcmcDiag Class Definition}
\usage{
\method{print}{mcmcDiag}(x)
}
\arguments{
\item{x}{Object of class mcmcDiag}
}
\value{
Invisible NULL, prints to console
}
\description{
A class for storing and handling MCMC diagnostics.
}
\details{
slot diagnostics A list containing MCMC diagnostic information.
slot transformedDraws A list containing transformed MCMC draws.
slot call A list containing information about the function call.

docType class
name mcmcDiag-class
rdname mcmcDiag-class
exportClass mcmcDiag
Print Method for mcmcDiag Objects

Prints MCMC diagnostics and relevant information.

param x An object of class \code{mcmcDiag}.

return Prints MCMC diagnostics to the console.

exportMethod print
aliases print.mcmcDiag
Print method for mcmcDiag objects
}
