% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag-sgelmanRubin.R
\name{sgelmanRubin}
\alias{sgelmanRubin}
\title{Calculate the Gelman-Rubin diagnostic of draws from an MCMC algorithm}
\usage{
sgelmanRubin(mhDraws, ...)
}
\arguments{
\item{mhDraws}{List of data.frames with two columns. Each data.frame represents a single chain. Column names should be val.1 (for values) and t (for chain iteration).}

\item{...}{Catches unnecessary additional arguments}
}
\value{
Data.frame with 1 row and 2 columns. First entry gives estimated GR statistic, second gives upper 95\% limit for GR statistic.
}
\description{
Calculate the Gelman-Rubin diagnostic of draws from an MCMC algorithm
}
