% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf-standardTransform.R
\name{standardTransform}
\alias{standardTransform}
\title{Transforms an mcmcObj into a dataframe with no modifications to values}
\usage{
standardTransform(mhDraws, ...)
}
\arguments{
\item{mhDraws}{An object of class mcmcObj}

\item{...}{Not used.}
}
\value{
data.frame with rows that represent MCMC draws. Columns named val.1, val.2,...
contain different dimensions of the draws, column named Posterior contains (non-normalized)
posterior value of the draw
}
\description{
Transforms an mcmcObj into a dataframe with no modifications to values
}
