% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag-straceplot.R
\name{straceplot}
\alias{straceplot}
\title{Generate a traceplot of draws from a multi-chain MCMC}
\usage{
straceplot(mhDraws, method = NULL, ...)
}
\arguments{
\item{mhDraws}{List of data.frames with two columns. Each data.frame represents a single chain. Column names should be val.1 (for values) and t (for chain iteration).}

\item{method}{Character string - Name of method used to generate traceplot. Is used to generate the title of the traceplot.}

\item{...}{Catches unused arguments}
}
\value{
ggplot2 plot object showing traceplot
}
\description{
Generate a traceplot of draws from a multi-chain MCMC
}
