% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf-sumTransform.R
\name{sumTransform}
\alias{sumTransform}
\title{Transforms a list of mcmcObjs into a list of data.frames using the 'sum' transformation}
\usage{
sumTransform(mhDraws, ...)
}
\arguments{
\item{mhDraws}{A list of mcmcObjs}

\item{...}{Catches extra arguments. Not used.}
}
\value{
List of data.frames with columns 'val.1' which is the sum of all dimensions of each MCMC
draw, 'Posterior' which is the (non-normalized) posterior value of each MCMC draw and 't'
which gives the within-chain ordering of the MCMC draws. Each data.frame is a separate chain.
}
\description{
Transforms a list of mcmcObjs into a list of data.frames using the 'sum' transformation
}
